Added view
Map tiles have size min(windowWidth,windowHeight)/n where n is the size of the map

Assets have a diameter, this diameter is a factor of the tile Size. So an asset with a diameter of 0.8 will become a circle with diameter 80% * mapTileSize

The view automatically resizes when the window is resized.

Running the main will show an example